Safari-611.1.10/LayoutTests/fast/events/scale-and-scroll-body-expected.txt

PASS window.document.scrollingElement.scrollTop is 30
PASS window.document.scrollingElement.scrollLeft is 30
PASS window.document.scrollingElement.scrollTop is 100
PASS window.document.scrollingElement.scrollLeft is 100
PASS window.scrollX is 100
PASS window.scrollY is 100
